新增编辑器键盘收起
